+++ backend/src/MonChenil.Api/Program.cs
@@ -3,6 +3,8 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using MonChenil.Data;
 using MonChenil.Domain.Models;
+using MonChenil.Domain.Users;
+using MonChenil.Domain.Pets;
 using MonChenil.Infrastructure.Repositories;
 
 var builder = WebApplication.CreateBuilder(args);
