+++ backend/src/MonChenil.Domain/Models/TimeSlot.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Security.Principal;
 using System.Text;
 using System.Threading.Tasks;
+using MonChenil.Domain.Pets;
 
 namespace MonChenil.Domain.Models
 {
