+++ backend/src/MonChenil.Domain/Pets/Dog.cs
@@ -1,0 +1,11 @@
+﻿using MonChenil.Domain.Users;
+
+namespace MonChenil.Domain.Pets;
+
+public class Dog : Pet
+{
+    public Dog(string name, User owner) : base(name, owner)
+    {
+        Type = PetType.Dog;
+    }
+}
