+++ backend/src/MonChenil.Domain/Pets/Pet.cs
@@ -1,0 +1,17 @@
+﻿using MonChenil.Domain.Users;
+
+namespace MonChenil.Domain.Pets;
+
+public abstract class Pet
+{
+    protected string Name;
+    protected PetType Type;
+    protected User Owner;
+
+    public Pet(string name, User owner)
+    {
+        Name = name;
+        Owner = owner;
+    }
+}
+
