+++ backend/src/MonChenil.Domain/Pets/PetType.cs
@@ -1,0 +1,8 @@
+﻿namespace MonChenil.Domain.Pets;
+
+public enum PetType
+{
+    Dog,
+    Cat,
+}
+
