+++ backend/src/MonChenil.Infrastructure/Repositories/PetRepository.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text;
-using System.Threading.Tasks;
-using MonChenil.Domain.Models;
+﻿using System.Linq.Expressions;
+using MonChenil.Domain.Pets;
 
 namespace MonChenil.Infrastructure.Repositories
 {
